Renaming hapi test files and filepaths containing 17 and 18 to 20. Also updating some 17-containing test names to remove version.
