+++ lib/instrumentation/hapi.js
@@ -27,7 +27,7 @@ module.exports = function initialize(agent, hapi, moduleName, shim) {
   } else if (hapi.Server) {
     // Server.connection was removed in v17
     if (!hapi.Server.prototype.connection) {
-      return require('./hapi/hapi-17')(agent, hapi, moduleName, shim)
+      return require('./hapi/hapi-20')(agent, hapi, moduleName, shim)
     }
   }
 }
