+++ test/versioned/hapi/hapi-v20/hapi-20-utils.js
@@ -14,7 +14,7 @@ exports.getServer = function getServer(cfg) {
   const opts = cfg.options || {}
   const hapi = cfg.hapi || require('@hapi/hapi')
 
-  // v17 exports two references to the server object,
+  // v17 and later exports two references to the server object,
   // so we'll let fate decide which to use for a given test
   const servers = ['Server', 'server']
   const server = servers[Math.round(Math.random())]
