+++ test/versioned/hapi/hapi-v20/newrelic.js
