+++ test/versioned/hapi/hapi-v20/plugins.tap.js
@@ -8,7 +8,7 @@
 const helper = require('../../../lib/agent_helper')
 const request = require('request')
 const tap = require('tap')
-const utils = require('./hapi-18-utils')
+const utils = require('./hapi-20-utils')
 
 tap.test('Hapi Plugins', function (t) {
   t.autoend()
